Update reader
The island now does not get a char asociated with because we run
out of them for many islands. now we jst count up and give the
island a number.

Added todo for asymetric maps that are currently not suported
and provided a qwuck fix for the normaly 2 provided gridsize
variables in the first line.